IronSky server

+++ servers/IronSky/start.sh
@@ -8,8 +8,10 @@ MC_PROJECT_NAME=${MC_PROJECT_NAME-$(basename $(pwd))}
 
 cd config && rsync -a --exclude=exclude.txt --exclude-from='exclude.txt' ../../../clients/${MC_PROJECT_NAME}/config/config/ ./; cd -
 cd mods && rsync --delete -a --exclude=exclude.txt --exclude-from='exclude.txt' ../../../clients/${MC_PROJECT_NAME}/mods/ ./; cd -
-mkdir -p worlds/world/datapacks
-cd worlds/world/datapacks && rsync --delete -a ../../../datapacks/ ./; cd -
+if [[ -d worlds/world/datapacks ]]; then
+  mkdir -p worlds/world/datapacks
+  cd worlds/world/datapacks && rsync --delete -a ../../../datapacks/ ./; cd -
+fi
 
 mc-docker-run --name "${MC_PROJECT_NAME}" --data "${MC_DATA_DIR-$(pwd)}" --port ${MC_PORT} \
 --image "${CONTAINER_IMAGE}" \
